requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
boto3==1.34.140
botocore==1.34.140
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.2
yarl==1.9.4
